add ability to upload image
- fix profile picture
- use profile picture of a signed in user
- separate api calls from actions creators
[Starts #171896067}
