+++ package.json
@@ -16,6 +16,9 @@
     "babel-preset-react-native": "^4.0.1",
     "expo": "^35.0.0",
     "expo-app-auth": "~7.0.0",
+    "expo-constants": "~7.0.0",
+    "expo-image-picker": "~7.0.0",
+    "expo-permissions": "~7.0.0",
     "fs": "^0.0.1-security",
     "graphql": "^14.5.3",
     "graphql-tag": "^2.10.1",
