+++ src/api/queries/authenticationQueries.js
@@ -1,0 +1,9 @@
+import gql from "graphql-tag";
+
+export default gql`
+  mutation createUser($accessToken: String) {
+    createUser(accessToken: $accessToken) {
+      token
+    }
+  }
+`;
