+++ src/screens/DrawerScreen/index.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import {
   Text,
   View,
@@ -7,10 +7,25 @@ import {
   AsyncStorage
 } from "react-native";
 import { DrawerNavigatorItems } from "react-navigation-drawer";
-import profileAvatar from "../../assets/profile.jpg";
 import styles from "./styles";
 
+const getProfileImage = async () => {
+  const profilePicture = await AsyncStorage.getItem("userProfile");
+  const profile = await JSON.parse(profilePicture);
+  return profile;
+};
+
 export default props => {
+  const [profile, setProfile] = useState({});
+  const fetchProfile = async () => {
+    const userProfile = await getProfileImage();
+    setProfile(userProfile);
+  };
+
+  useEffect(() => {
+    fetchProfile();
+  }, []);
+
   const logOut = async () => {
     await AsyncStorage.removeItem("token");
     props.navigation.navigate("Login");
@@ -18,8 +33,13 @@ export default props => {
   return (
     <View style={styles.container}>
       <View style={styles.avatarContainer}>
-        <Image source={profileAvatar} style={styles.profileAvatar} />
-        <Text style={styles.profileText}>Yves Iraguha</Text>
+        <Image
+          source={{ uri: profile ? profile.picture : null }}
+          style={styles.profileAvatar}
+        />
+        <Text
+          style={styles.profileText}
+        >{`${`${profile.family_name} ${profile.given_name}`} `}</Text>
       </View>
       <View>
         <DrawerNavigatorItems {...props} />
