+++ src/screens/Expenses/EditScreen/Components/SaveButton.js
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 import { Button } from "react-native";
 import styles from "./styles";
-import { handleEditingExpense } from "../../../../redux/createExpense";
+import { handleEditingExpense } from "../../../../redux/actionsCreators/expenses";
 
 const SaveButton = props => {
   const {
