+++ src/screens/Expenses/components/HeaderRight.js
@@ -1,4 +1,4 @@
-import React,  from "react";
+import React from "react";
 import { View, Image } from "react-native";
 import profilePic from "../../../assets/profile.jpg";
