+++ src/screens/Expenses/index.js
@@ -1,14 +1,15 @@
 /* eslint-disable no-nested-ternary */
-import React, { useEffect } from "react";
+import React, { useEffect, useLayoutEffect, useState } from "react";
 import {
   View,
   SectionList,
   Text,
   ActivityIndicator,
-  Image
+  Image,
+  AsyncStorage
 } from "react-native";
 import { connect } from "react-redux";
-import { fetchExpenses } from "../../redux/actionsCreators";
+import { fetchExpenses } from "../../redux/actionsCreators/expenses";
 import styles from "./styles";
 import Expense from "./components/Expense";
 import normalizeData from "../../helpers/normilizeData";
@@ -17,11 +18,33 @@ import HeaderLeft from "./components/HeaderLeft";
 import noData from "../../assets/undraw_empty_xct9.png";
 import NewExpenseModal from "./components/NewExpenseModal";
 
+const getProfileImage = async () => {
+  const profilePicture = await AsyncStorage.getItem("userProfile");
+  const profile = await JSON.parse(profilePicture);
+  return profile;
+};
+
 const Home = props => {
+  const [profileAvatar, setProfileAvatar] = useState(null);
+
+  const fetchProfile = async () => {
+    const { picture } = await getProfileImage();
+    setProfileAvatar(picture);
+  };
+
   useEffect(() => {
     const { loadExpenses } = props;
     loadExpenses();
   }, []);
+  useEffect(() => {
+    fetchProfile();
+  });
+
+  useLayoutEffect(() => {
+    const { navigation } = props;
+    navigation.setParams({ profileAvatar });
+  }, [profileAvatar]);
+
   const { expenses, apiInProgress, navigation } = props;
 
   return (
