+++ src/screens/Investments/EditScreen/Components/SaveButton.js
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 import { Button } from "react-native";
 import styles from "./styles";
-import { handleEditingInvestment } from "../../../../redux/createExpense";
+import { handleEditingInvestment } from "../../../../redux/actionsCreators/investments";
 
 const SaveButton = props => {
   const {
