+++ src/screens/Investments/NewInvestment/index.js
@@ -7,6 +7,9 @@ import {
   Platform
 } from "react-native";
 import { EvilIcons } from "@expo/vector-icons";
+import * as ImagePicker from "expo-image-picker";
+import Constants from "expo-constants";
+import * as Permissions from "expo-permissions";
 import imageUrl from "../../../assets/expense.jpeg";
 import Item from "./Components/ExpenseProperty";
 import styles from "./styles";
@@ -20,6 +23,7 @@ const NewInvestmentScreen = ({ navigation }) => {
   const [keyboardHeight, setKeyboardHeight] = useState(0);
   const [imageHeight, setImageHeight] = useState(IMAGE_HEIGHT);
   const [investment, setInvestment] = useState({});
+  const [selectedImage, setSelectedImage] = useState(null);
   const onInputChange = (field, value) => {
     setInvestment({ ...investment, [field]: value });
   };
@@ -34,6 +38,30 @@ const NewInvestmentScreen = ({ navigation }) => {
     setImageHeight(IMAGE_HEIGHT);
   };
 
+  const getPermissionAsync = async () => {
+    if (Constants.platform.ios) {
+      const { status } = await Permissions.askAsync(Permissions.CAMERA_ROLL);
+      if (status !== "granted") {
+        alert("Sorry, we need camera roll permissions to upload images!");
+      }
+    }
+  };
+
+  const pickImage = async () => {
+    const result = await ImagePicker.launchImageLibraryAsync({
+      mediaTypes: ImagePicker.MediaTypeOptions.Images,
+      allowsEditing: true,
+      aspect: [1, 1],
+      quality: 1,
+      base64: true
+    });
+    if (result.cancelled) {
+      return;
+    }
+    const base64Img = `data:image/jpg;base64,${result.base64}`;
+    setSelectedImage(result.uri);
+    onInputChange("image", base64Img);
+  };
   useEffect(() => {
     const firstEvent =
       Platform.OS === "ios" ? "keyboardWillShow" : "keyboardDidShow";
@@ -53,6 +81,9 @@ const NewInvestmentScreen = ({ navigation }) => {
     };
   }, []);
 
+  useEffect(() => {
+    getPermissionAsync();
+  });
   useLayoutEffect(() => {
     navigation.setParams(investment);
   }, [investment]);
@@ -60,11 +91,11 @@ const NewInvestmentScreen = ({ navigation }) => {
   return (
     <View style={{ paddingBottom: keyboardHeight }}>
       <ImageBackground
-        source={imageUrl}
+        source={selectedImage ? { uri: selectedImage } : imageUrl}
         imageStyle={styles.imageStyle}
         style={[styles.expenseImage, { height: imageHeight }]}
       >
-        <TouchableOpacity style={styles.editButton}>
+        <TouchableOpacity style={styles.editButton} onPress={() => pickImage()}>
           <EvilIcons name="camera" size={56} color="white" />
         </TouchableOpacity>
       </ImageBackground>
@@ -101,15 +132,6 @@ const NewInvestmentScreen = ({ navigation }) => {
           value={investment.matureDate}
           onTextChange={onInputChange}
           editable={false}
-        />
-
-        <Item
-          title="Location"
-          name="location"
-          iconName="location-on"
-          placeholder="Location"
-          value={investment.quantity}
-          onTextChange={onInputChange}
         />
       </View>
     </View>
