+++ src/screens/Investments/components/HeaderLeft.js
@@ -1,14 +1,19 @@
 import React from "react";
 import { TouchableOpacity, Image } from "react-native";
-import profileAvatar from "../../../assets/profile.jpg";
+
 import styles from "./styles";
 
-export default ({ navigation }) => (
-  <TouchableOpacity onPress={navigation.openDrawer}>
-    <Image
-      source={profileAvatar}
-      resizeMode="contain"
-      style={styles.profileAvatar}
-    />
-  </TouchableOpacity>
-);
+export default ({ navigation }) => {
+  const {
+    state: { params }
+  } = navigation;
+  return (
+    <TouchableOpacity onPress={navigation.openDrawer}>
+      <Image
+        source={{ uri: params ? params.profileAvatar : null }}
+        resizeMode="cover"
+        style={styles.profileAvatar}
+      />
+    </TouchableOpacity>
+  );
+};
