+++ src/screens/login/index.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { View, Image, ActivityIndicator, AsyncStorage } from "react-native";
 import Toast, { DURATION } from "react-native-easy-toast";
 import { connect } from "react-redux";
-import { authenticateUser } from "../../redux/actionsCreators";
+import registerUser from "../../redux/actionsCreators/authentication";
 import LoginButton from "./components/LoginButton";
 
 import personalFinance from "../../assets/undraw_personal_finance_tqcd.png";
@@ -29,7 +29,7 @@ class LoginScreen extends Component {
   };
 
   handleAuthentication = () => {
-    const { authenticateUser, navigation } = this.props;
+    const { registerUser: authenticateUser, navigation } = this.props;
     authenticateUser(navigation);
   };
 
@@ -91,5 +91,5 @@ const mapStateToProps = state => ({
 });
 export default connect(
   mapStateToProps,
-  { authenticateUser }
+  { registerUser }
 )(LoginScreen);
